for plyfitting class: changed some fo the thresholds fo band quality scores and which orders of polynomials these bands are allowed to explore to help suit ASAS-SN-17jz, 18jd and PS1-10adi (they also seem to work fine for Phil's ANTs so all good)
Also made the 'gapsize' an input into the polyfitting class so the user can choose what span of MJD is considered a gap in the data, a region over which we will not interpolate at ALL because the polyfits vary too wildly in these regions.

Also made the band marker dict an input into the polyfitting class.

polyfitting function: changed the legend on the plotting function to allow for all 23 bands of ASASSN18jd to fit lol.

Also added a function which plots the poltfits as a subplot where each subplot is an individual band so you can see them one by one without them overlapping on one another. Gets very crowded for ASASSN18jd. This will be plotted whenever plot_polyfit is input as True