Added a function to convert from flux density to magnitude and vice versa.
Did the host subtraction on ASASSN-18jd and made sure that I didn't mess up the data in doing so lol.

For ASASSN-17jz, I just changed how they document MJD, since the data that we are given is binned. However, I checked the bin sizes and they were almost exclusively <1 day in MJD, so I approximate that the measurements provided within this bin are all taken at the central MJD of this bin to allow for further analysis.

Changed PS1-10adi and PS1-13jw's bands from 'Vega?' to '(Vega)' because I KNOW they are Vega now lol

Also added a bunch of code in an attempt to convert the Vega mags to AB mags which I found errors in. ALlso can't convert to app mag until we have the SED to (reverse) K-correct the mags. Also wrote some code to calculate the rest frame luminosity of PS1-10adi which does actually work (I think) but I realised that I don't need to do this, I can just do it when we cluclate the rest frame luminosity of all of the other ANTs by inputting m = Vega abs mag, d = 10pc and therefore approximate that z = 0.0, then input the band's zerppoint in the Vega system.