in polyfit lightcurve class, I added a dictionary input which allows you to inputmanually identified stragglers for each ANT (although assumes only 1 per band so if you have more than that will need to update it slightly).
Also corrected the calculation of rest frame days since peak.

Also forced the calculation of peak MJD to occur within the first 60% of teh light curve, this was because in one of the ANTs, the polyfit flies off in a gap in the light curve, wuite far on the decline. This meant that the peka was being taken to be way out on the decline which is just wrong.