Commented out a few things to get us closer for the DBB to using the curve_fit model param, then using the brute force gridding to get the uncertainties on this parameter, so getting rid of storing the brute force grididng's 'optimal' parameter. I don't save these since the grid is too coarse to get an accurate value.
When it comes to sampling parameter values frmo the uncertainties, we make sure not to smaple beyond the parameter seach limits, like the regular bounds for the SED fit or the UVOT-gudied ones.

Deleted the DBB_brute_param_search_lims because I realised this limits the search too early, we want to do the brute grid on the entire +/- 2sigma to get accurate uncertainties, then limit the region over which we sample from.

Hopefully successfully implemented the double_BB_curvefit_then_brute()