Made changes to the histogram plotting BUT HAVEN'T RAN THE CODE YET
We now skip fitting and plotting the DBB histograms for ANTs which don't have much UV data. We also change the way we find the UV MJDs to be more accurate, by filtering by a wavelength threshold to define what is and is not UV, rather than the observed band.